Move flag app out of the apps directory.
