+++ mapstory/flag/templates/flag/flag_form.html
