+++ mapstory/flag/templates/flag/thank_you.html
