+++ mapstory/flag/templatetags/flag_tags.py
