+++ mapstory/flag/urls.py
@@ -4,6 +4,6 @@ from django.views.generic import TemplateView
 
 
 urlpatterns = patterns("",
-    url(r"^$", "mapstory.apps.flag.views.flag", name="flag"),
+    url(r"^$", "mapstory.flag.views.flag", name="flag"),
     url(r'^thank_you', TemplateView.as_view(template_name="flag/thank_you.html"), name='flag-reported'),
 )
