+++ mapstory/settings/base.py
@@ -130,7 +130,7 @@ if is_valid(os.getenv("ALLAUTH_GEOAXIS_HOST")):
 
 MAPSTORY_APPS = (
 
-    'mapstory.apps.flag', # - temporarily using this instead of the flag app for django because they need to use AUTH_USER_MODEL
+    'mapstory.flag', # - temporarily using this instead of the flag app for django because they need to use AUTH_USER_MODEL
 
 )
