+++ mapstory/social_settings.py
@@ -3,7 +3,7 @@ from django.db.models import Q
 from geonode.people.models import Profile
 from mailer import send_html_mail
 
-from mapstory.apps.flag.signals import content_flagged
+from mapstory.flag.signals import content_flagged
 
 
 def flag_handler(flagged_instance, flagged_content, **kw):
